% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{addClassAttribute,ANY,character-method}
\alias{addClassAttribute,ANY,character-method}
\title{Add Class Attribute}
\usage{
\S4method{addClassAttribute}{ANY,character}(obj, class_name)
}
\arguments{
\item{obj}{\code{\link{ANY}}.}

\item{class_name}{\code{\link{character}}.}
}
\value{
Object with updated class graph (\code{c(<class_name>, <class>)})
		with the first entry corresponding to \code{class_name}.
}
\description{
See \code{\link[filesystr]{addClassAttribute}}.
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{addClassAttribute}
}
}

