% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addClassAttribute}
\alias{addClassAttribute}
\title{Add Class Attribute}
\usage{
addClassAttribute(obj, class_name)
}
\arguments{
\item{obj}{\strong{Signature argument}.
Object to be updated.}

\item{class_name}{\strong{Signature argument}.
Object containing class name information.}
}
\description{
Adds class attribute to an object and returns it.
}
\examples{
\dontrun{

obj <- "hello world!"
obj_new <- filesystr::addClassAttribute(
  obj = obj, 
  class_name = "RappFilesystemCopyResultS3"
) 
obj_new
class(obj_new)
filesystr::addClassAttribute(
  obj = obj, 
  class_name = "RappFilesystemCopyResultsS3"
)
filesystr::addClassAttribute(
  obj = obj, 
  class_name = "RappFilesystemDecomposedPathS3"
)
filesystr::addClassAttribute(
  obj = obj, 
  class_name = "RappFilesystemDirectoryS3"
)
filesystr::addClassAttribute(
  obj = obj, 
  class_name = "RappFilesystemFileS3"
)

## Condition handling //
try(filesystr::addClassAttribute(obj = obj, 
  class_name = "NonExistingClass"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{addClassAttribute-character-method}
}
}

