% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{asDecomposedPath,character-method}
\alias{asDecomposedPath,character-method}
\title{As Decomposed Path}
\usage{
\S4method{asDecomposedPath}{character}(path, ensure, ...)
}
\arguments{
\item{path}{\code{\link{character}}.}

\item{...}{Further arguments of:
\code{\link[filesystr]{decomposePath}}.}
}
\value{
\code{FilePath.S3}. Identical to \code{path} with
   updated class table.
}
\description{
See \code{\link[base]{asDecomposedPath}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[base]{asDecomposedPath}
}
}

