% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{asDecomposedPath,RappFilesystemDecomposedPathS3-method}
\alias{asDecomposedPath,RappFilesystemDecomposedPathS3-method}
\title{As Decomposed Path}
\usage{
\S4method{asDecomposedPath}{RappFilesystemDecomposedPathS3}(path = ".", ...)
}
\arguments{
\item{path}{\code{\link{RappFilesystemDecomposedPathS3}}.}

\item{...}{Further arguments of:
\code{\link[filesystr]{decomposePath}}.}
}
\value{
See method
   \code{\link{asDecomposedPath-character-method}}.
}
\description{
See \code{\link[base]{asDecomposedPath}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
   \link[base]{asDecomposedPath},
   \link[base]{asDecomposedPath-character-method}
}
}

