% Generated by roxygen2 (4.0.2): do not edit by hand
\name{asDecomposedPath}
\alias{asDecomposedPath}
\title{As Decomposed Path}
\usage{
asDecomposedPath(path = ".", ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{...}{Further arguments of:
\code{\link[filesystr]{decomposePath}}.}
}
\description{
Coerces files system ressources' paths into objects of class
\code{DecomposedPath.S3}.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[base]{asDecomposedPath-character-method}
}
}

