% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{asFilePath,character-method}
\alias{asFilePath,character-method}
\title{As File}
\usage{
\S4method{asFilePath}{character}(path = ".", ensure = FALSE, ...)
}
\arguments{
\item{path}{\code{\link{character}}.}

\item{ensure}{\code{logical}.
Ensure file existence (\code{TRUE}) or not (\code{FALSE}, default).}

\item{...}{Further arguments of:
\code{\link[filesystr]{asFilePath}}.}
}
\value{
\code{File.S3}. Identical to \code{path} with
   updated class table.
}
\description{
See \code{\link[base]{asFilePath}}
}
\examples{
\dontrun{

## Auxiliary function //
.cleanTempDir <- function(x) {
  if (grepl(basename(tempdir()), x)) {
    unlink(x, recursive = TRUE, force = TRUE)
  }
}

## Path //
path_0 <- file.path(tempdir(), "path/test.txt")

## character //
res <- asFilePath(path = path_0)
res
class(res)

## File.S3 //
res <- asFilePath(path = asFilePath(path_0))
res
class(res)

## Ensure //
file.exists(asFilePath(path = path_0, ensure = TRUE))
.cleanTempDir(x = path_0)

file.exists(
  asFilePath(path = asFilePath(path_0), ensure = TRUE)
)
.cleanTempDir(x = path_0)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
   \link[base]{asFilePath}
}
}

