% Generated by roxygen2 (4.0.1): do not edit by hand
\name{asFilePath}
\alias{asFilePath}
\title{As File Path}
\usage{
asFilePath(path = ".", ensure = FALSE, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing file path information.}

\item{ensure}{\code{logical}.
Ensure file existence (\code{TRUE}) or not (\code{FALSE}, default).}

\item{...}{Further arguments of:
\code{\link[filesystr]{asFilePath}}.}
}
\description{
Turns a \code{character} object denoting a file path into an object
of class \code{File.S3}.
}
\examples{
\dontrun{

## Auxiliary function //
.cleanTempDir <- function(x) {
  if (grepl(basename(tempdir()), x)) {
    unlink(x, recursive = TRUE, force = TRUE)
  }
}

## Path //
path_0 <- file.path(tempdir(), "path/test.txt")

## character //
res <- asFilePath(path = path_0)
res
class(res)

## File.S3 //
res <- asFilePath(path = asFilePath(path_0))
res
class(res)

## Ensure //
file.exists(asFilePath(path = path_0, ensure = TRUE))
.cleanTempDir(x = path_0)

file.exists(
  asFilePath(path = asFilePath(path_0), ensure = TRUE)
)
.cleanTempDir(x = path_0)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
   \link[base]{asFilePath-character-method}
}
}

