% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{conditionalDelete,character,character-method}
\alias{conditionalDelete,character,character-method}
\title{Conditional Delete}
\usage{
\S4method{conditionalDelete}{character,character}(path, condition,
  content_only = FALSE, ...)
}
\arguments{
\item{path}{\code{\link{character}}.}

\item{condition}{\code{\link{character}}.}

\item{content_only}{\code{\link{logical}}.
Only relevant if \code{path} corresponds to a directory path.
\code{TRUE}: only delete directory content, not the directory itself;
\code{FALSE}: delete entire directory.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[filesystr]{conditionalDelete-character-character-method}}
}
\description{
See \code{\link[base]{conditionalDelete}}
}
\examples{
\dontrun{

## Example content //
path <- file.path(tempdir(), "conditionalDelete/dir")    
dir.create(path, recursive = TRUE, showWarnings = FALSE)
file.create(file.path(path, paste0("file", 1:3)))

## Delete file if its in a directory that is a subdirectory of 
## the temp directory //
fpath <- file.path(path, "file1")
conditionalDelete(path = fpath, condition = basename(tempdir()))
file.exists(fpath)
file.exists(path)

## When condition is not met //
fpath <- file.path(path, "file2")
conditionalDelete(path = fpath, condition = "abcd")
file.exists(fpath)

## Delete *content* of directory if its a subdirectory of the temp directory //
conditionalDelete(path = path, condition = basename(tempdir()), 
  content_only = TRUE)
file.exists(path)

## Delete directory if its a subdirectory of the temp directory //
conditionalDelete(path = path, condition = basename(tempdir()))
file.exists(path)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[base]{conditionalDelete}
}
}

