% Generated by roxygen2 (4.0.1): do not edit by hand
\name{copyDirectory}
\alias{copyDirectory}
\title{Copy Directory/Directories}
\usage{
copyDirectory(from, to, pattern, ensure_to = FALSE, overwrite = FALSE,
  recursive = TRUE, ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing source directory information.}

\item{to}{\strong{Signature argument}.
Object containing target directory information.}

\item{pattern}{\strong{Signature argument}.
Filter for matching directory specific directory content.}

\item{ensure_to}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means non-existing target directories in \code{to} are
created, \code{FALSE} means they will trigger an error.}

\item{overwrite}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means existing directories are overwritten,
\code{FALSE} means they are not.}

\item{recursive}{\code{\link{logical}}.
\code{TRUE} means subdirectories and their content are copied,
\code{FALSE} means they are not.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Copies the contents of a directory, possibly recursively.
}
\note{
Target directories that don't exist are created, silently (assuming
write permission).
}
\examples{
\dontrun{

## Directory preparations //
to <- file.path(tempdir(), "copyDirectory", c("to1", "to2"))
sapply(to, dir.create, recursive=TRUE, showWarnings=FALSE)

## Utility functions //
.openRessource <- function(dir = getwd()) {
  if (.Platform['OS.type'] == "windows"){
    shell.exec(dir)
  } else {
    system(paste(Sys.getenv("R_BROWSER"), dir))
  }
}
.carefulCleanup <- function(x, pattern=basename(tempdir()), content.only=TRUE) {
  out <- sapply(x, function(ii) {
      out <- FALSE
      if (grepl(pattern, ii)) {
        out <- !as.logical(unlink(ii, recursive=TRUE, force=TRUE))
        if (out & content.only) {
          dir.create(ii, recursive=TRUE, showWarnings=FALSE)    
        }
      }        
      out
    })
  invisible(out)
}

## Open directories in file system browser //
.openRessource(to[1])
.openRessource(to[2])

## One source, one target //
copyDirectory(from = R.home("etc"), to = to[1])
.carefulCleanup(to[1])

## One source, two targets //
copyDirectory(from = R.home("etc"), to = to)
.carefulCleanup(to)

## Two sources, one target //
## Combines content of 'etc' and 'tests'. If there would be any conflicts,
## the would result in an error as 'overwrite = FALSE'.
## Otherwise, existing content is overwritten.
copyDirectory(from = c(R.home("etc"), R.home("tests")), to = to[1])
.carefulCleanup(to)

## Two sources, two targets //
copyDirectory(from = c(R.home("etc"), R.home("tests")), to = to)
.carefulCleanup(to)

## Non-recursive //
## Only files directly below 'from' and actual subdirectories are copied, 
## *NOT* their content.
copyDirectory(from = R.home("etc"), to = to[1], recursive = FALSE)

## Subsequent recursive copy, existing content overwritten //
## This will add the actual content of the subdirectories.
copyDirectory(from = R.home("etc"), to = to[1], overwrite = TRUE)
.carefulCleanup(to[1])

## Condition handling //
## Handling of non-existing target directories:
.carefulCleanup(to, content.only = FALSE)
try(copyDirectory(from = R.home("etc"), to = to[1]))
## Existence of target root directories can be ensured:
copyDirectory(from = R.home("etc"), to = to[1], ensure_to =TRUE)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{copyDirectory-character-character-character-method}
}
}

