% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DecomposedPath.S3}
\alias{DecomposedPath.S3}
\title{Class: DecomposedPath.S3}
\usage{
DecomposedPath.S3(.x, decomposed = data.frame())
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
attribute should be updated so that it becomes an instance of class
\code{DecomposedPath.S3}. Mainly intended for rapid prototyping
purposes}
}
\value{
Instance of class \code{DecomposedPath.S3}.
}
\description{
Directory class (S3).
}
\section{Fields}{

\describe{
\item{\code{decomposed}}{\code{\link{data.frame}}. Decomposed path(s).}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
DecomposedPath.S3(data.frame(
  directory=tempdir(), 
  filename=basename(tempfile()), 
  extension = "")
)  
DecomposedPath.S3("anything")  

## Formal use (explicitly using 'fields') //
DecomposedPath.S3()
DecomposedPath.S3(decomposed = data.frame(
  directory=tempdir(), 
  filename=basename(tempfile()), 
  extension = "")
)

## Recommended: inlcude namespace //
filesystr::DecomposedPath.S3(decomposed = data.frame(
  directory=tempdir(), 
  filename=basename(tempfile()), 
  extension = "")
)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[filesystr]{File.S3},
  	\link[filesystr]{Directory.S3}
}
}

