% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DirectoryCopyResult.S3}
\alias{DirectoryCopyResult.S3}
\title{Class: DirectoryCopyResult.S3}
\usage{
DirectoryCopyResult.S3(.x, from = character(), to = character(),
  status = numeric())
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
attribute should be updated so that it becomes an instance of class
\code{DirectoryCopyResult.S3}. Mainly intended for rapid prototyping
purposes}
}
\value{
Instance of class \code{DirectoryCopyResult.S3}.
}
\description{
Copy result class (S3).
}
\section{Fields}{

\describe{
\item{\code{from}}{\code{\link{character}}. Copy source location.}

\item{\code{to}}{\code{\link{character}}. Copy target location.}

\item{\code{status}}{\code{\link{numeric}}. Copy result.
\itemize{
   \item{\code{0}: }{Failure}
   \item{\code{1}: }{Success}
}}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
DirectoryCopyResult.S3(1)  
DirectoryCopyResult.S3(rep(1, 3))  
DirectoryCopyResult.S3(as.list(rep(1, 3)))
DirectoryCopyResult.S3(TRUE)  

## Formal use (explicitly using 'fields') //
DirectoryCopyResult.S3()
DirectoryCopyResult.S3(status = 1)
DirectoryCopyResult.S3(status = rep(1, 3))
DirectoryCopyResult.S3(status = as.list(rep(1, 3)))

## Recommended: inlcude namespace //
filesystr::DirectoryCopyResult.S3(1)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
  	\link[DirectoryCopyResultsystr]{File.S3}
}
}

