% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ensureDirectory,DirectoryPath.S3-method}
\alias{ensureDirectory,DirectoryPath.S3-method}
\title{Create Directory}
\usage{
\S4method{ensureDirectory}{DirectoryPath.S3}(path, strict, ...)
}
\arguments{
\item{path}{\code{\link{DirectoryPath.S3}}.}

\item{strict}{\code{\link{logical}}.
Should certain conditions trigger an error (\code{TRUE}) or simply result
in a return value of \code{FALSE} (\code{FALSE}, default).}

\item{...}{Further arguments to be passed to:
\code{\link[base]{dir.create}}.}
}
\value{
\code{\link{logical}} vector. \code{TRUE} for success, \code{FALSE}
		for failure.
}
\description{
See \code{\link[filesystr]{ensureDirectory}}.
}
\examples{
\dontrun{

## Auxiliary functions //
.carefulCleanup <- function(x, pattern=basename(tempdir())) {
  out <- sapply(x, function(ii) {
    out <- NA
    if (grepl(pattern, ii)) {
      out <- unlink(ii, recursive=TRUE, force=TRUE)
    }        
    out
  })
  invisible(out)
}

## Example content //
path <- file.path(tempdir(), "ensureDirectory")
sapply(path, dir.create, recursive=TRUE, showWarnings=FALSE)

path <- file.path(path, letters[1:3])
ensureDirectory(path = path)

## If directory already exists //
ensureDirectory(path = path)

## Clean up //
.carefulCleanup(x = path)
  

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
		\link[filesystr]{ensureDirectory}
}
}

