% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensureDirectory}
\alias{ensureDirectory}
\title{Create Directory}
\usage{
ensureDirectory(path, strict, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing directory information.}

\item{strict}{\code{\link{logical}}.
Should certain conditions trigger an error (\code{TRUE}) or simply result
in a return value of \code{FALSE} (\code{FALSE}, default).}

\item{...}{Further arguments to be passed to:
\code{\link[base]{dir.create}}.}
}
\description{
Creates directory/directories. Basically, an S4 interface to
\code{\link[base]{dir.create}}.
}
\examples{
\dontrun{

## Auxiliary functions //
.carefulCleanup <- function(x, pattern=basename(tempdir())) {
  out <- sapply(x, function(ii) {
    out <- NA
    if (grepl(pattern, ii)) {
      out <- unlink(ii, recursive=TRUE, force=TRUE)
    }        
    out
  })
  invisible(out)
}

## Example content //
path <- file.path(tempdir(), "ensureDirectory")
sapply(path, dir.create, recursive=TRUE, showWarnings=FALSE)

path <- file.path(path, letters[1:3])
ensureDirectory(path = path)

## If directory already exists //
ensureDirectory(path = path)

## Clean up //
.carefulCleanup(x = path)
  

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
  	\link[base]{ensureDirectory-character-method}
		\link[base]{dir.create}
}
}

