% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getDirectoryPath,character-method}
\alias{getDirectoryPath,character-method}
\title{Get Directory Path}
\usage{
\S4method{getDirectoryPath}{character}(path = ".", ...)
}
\arguments{
\item{path}{\code{\link{character}}. Vector of file paths.
Default: see \code{\link[filesystr]{getDirectoryPath-missing-RappCoreDirectorysys-RappCoreDirectorysys-method}}.}

\item{...}{Further arguments passed to:
\code{\link[filesystr]{decomposePath}}}
}
\value{
See method
   \link[filesystr]{getDirectoryPath-RappFilesystemDecomposedPathS3-method}
}
\description{
See generic: \code{\link[filesystr]{getDirectoryPath}}
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
getDirectoryPath(path = path)
getDirectoryPath(path = path, shortform = TRUE)
getDirectoryPath(path = character())
getDirectoryPath()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{getDirectoryPath},
  	\link[filesystr]{getDirectoryPath-RappFilesystemDecomposedPathS3-method}
}
}

