% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDirectoryPath}
\alias{getDirectoryPath}
\title{Get Directory Path}
\usage{
getDirectoryPath(path = ".", ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{...}{Further arguments passed to:
\code{\link[filesystr]{decomposePath}}}
}
\description{
Returns the directory path of paths of file system ressources.
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
getDirectoryPath(path = path)
getDirectoryPath(path = path, shortform = TRUE)
getDirectoryPath(path = character())
getDirectoryPath()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{getDirectoryPath-character-method}
}
}

