% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getFileExtension,character-method}
\alias{getFileExtension,character-method}
\title{Get File Extension}
\usage{
\S4method{getFileExtension}{character}(path = ".", ...)
}
\arguments{
\item{path}{\code{\link{character}}. Vector of file paths.
Default: see \code{\link[filesystr]{getFileExtension-missing-RappCoreFilesys-RappCoreFilesys-method}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \link[filesystr]{getFileExtension-RappFilesystemDecomposedPathS3-method}
}
\description{
See generic: \code{\link[filesystr]{getFileExtension}}
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
getFileExtension(path = path)
getFileExtension(path = character())
getFileExtension()

)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{getFileExtension},
  	\link[filesystr]{getFileExtension-RappFilesystemDecomposedPathS3-method}
}
}

