% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getFileName,character-method}
\alias{getFileName,character-method}
\title{Get File Name}
\usage{
\S4method{getFileName}{character}(path = ".", ...)
}
\arguments{
\item{path}{\code{\link{character}}. Vector of file paths.
Default: see \code{\link[filesystr]{getFileName-missing-RappCoreFilesys-RappCoreFilesys-method}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \link[filesystr]{getFileName-DecomposedPath.S3-method}
}
\description{
See generic: \code{\link[filesystr]{getFileName}}
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
getFileName(path = path)
getFileName(path = character())
getFileName()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
		\link[filesystr]{getFileName},
  	\link[filesystr]{getFileName-DecomposedPath.S3-method}
}
}

