% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getFileName,DecomposedPath.S3-method}
\alias{getFileName,DecomposedPath.S3-method}
\title{Get File Name}
\usage{
\S4method{getFileName}{DecomposedPath.S3}(path = ".", ...)
}
\arguments{
\item{path}{\code{\link{DecomposedPath.S3}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. Vector containing file names or
		\code{""} or \code{NA} for respective inputs (i.e. directory paths or
		\code{NA}).
}
\description{
See generic: \code{\link[filesystr]{getFileName}}
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
getFileName(path = path)
getFileName(path = character())
getFileName()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
		\link[filesystr]{getFileName},
		\link[filesystr]{getFileName-missing-method}
}
}

