% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFileName}
\alias{getFileName}
\title{Get File Name}
\usage{
getFileName(path = ".", ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Returns the file name of file paths.
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
getFileName(path = path)
getFileName(path = character())
getFileName()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
		\link[filesystr]{getFileName-character-method}
}
}

