% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getLockFileName,character-method}
\alias{getLockFileName,character-method}
\title{Get Name of Lock File}
\usage{
\S4method{getLockFileName}{character}(path, lock_suffix)
}
\arguments{
\item{path}{\code{\link{character}}. File path.}

\item{lock_suffix}{\code{\link{character}}.
Suffix to denote the lock dummy.}
}
\value{
\code{\link{character}}. Lock file name.
}
\description{
See generic: \code{\link[filesystr]{getLockFileName}}
}
\examples{
\dontrun{

## Non-existing file system ressource //
path <- file.path(tempdir(), "test.txt")
getLockFileName(path = path)

## Files //
file.create(path)
getLockFileName(path = path)

## Directories //
getLockFileName(path = tempdir())

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
   \link[filesystr]{getLockFileName}
}
}

