% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLockFileName}
\alias{getLockFileName}
\title{Get Name of Lock File}
\usage{
getLockFileName(path, lock_suffix = "_LOCK", ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing file path information.}

\item{lock_suffix}{\code{\link{character}}.
Suffix to denote the lock dummy.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Returns the name of the dummy file used for locking a file system ressource.
}
\examples{
\dontrun{

## Non-existing file system ressource //
path <- file.path(tempdir(), "test.txt")
getLockFileName(path = path)

## Files //
file.create(path)
getLockFileName(path = path)

## Directories //
getLockFileName(path = tempdir())

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[filesystr]{getLockFileName-character-method}
}
}

