% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getParentDirectory,character-method}
\alias{getParentDirectory,character-method}
\title{Get Path of Parent Directory (Vectorized)}
\usage{
\S4method{getParentDirectory}{character}(path, up = 1, ...)
}
\arguments{
\item{path}{\code{\link{character}}. Dimension: any.
Paths of file system ressources.}

\item{up}{\code{\link{numeric}}. Dimension: 1
Number of levels to go up (relative to the initial directory).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. A vector containing paths of parent
		directories based on \code{up}.
}
\description{
See generic: \code{\link[filesystr]{getParentDirectory}}.
}
\note{
The function is multi-vectorized: arguments \code{x} and \code{up}
(see \code{\link[base]{mapply}}.
}
\examples{
\dontrun{

path <- c(getwd(), R.home())

getParentDirectory(path = path)
getParentDirectory(path = path, up = 2)
getParentDirectory(path = path, up = 3)
getParentDirectory(path = path, up = c(1,4))

path <- "c:/Windows"
getParentDirectory(path = path, up = 2)
try(getParentDirectory(path = path, up = 3))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
		\link[filesystr]{getParentDirectory}
}
}

