% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isRessourceLocked}
\alias{isRessourceLocked}
\title{Is File System Ressource Locked}
\usage{
isRessourceLocked(path, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing file path information.}

\item{...}{Further arguments passed to subsequent functions/methods.
\code{\link[filesystr]{getLockFileName}}.}
}
\description{
Checks if a file is locked (\code{\link[filesystr]{isRessourceLocked}}.
}
\examples{
\dontrun{

## Create example content  
path <- file.path(tempdir(), "test/registry.rdata")
dir.create(dirname(path), showWarnings = FALSE)
file.create(path)

## Open directory in file system browser //
openRessource(path = dirname(path))

## File //
isRessourceLocked(path = path)
lockRessource(path = path)
## Note additional '*FILE_LOCK' file in directory 
isRessourceLocked(path = path)
unlockRessource(path = path)
## Note '*FILE_LOCK' file in directory has been removed again

## Directory //
isRessourceLocked(path = dirname(path))
lockRessource(path = dirname(path))
## Note additional '*DIR_LOCK' file in directory 
isRessourceLocked(path = dirname(path))
unlockRessource(path = dirname(path))
## Note '*DIR_LOCK' file in directory has been removed again

## Non-existing ressources //
try(isRessourceLocked(path = "abcd"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
  	\link[filesystr]{isRessourceLocked.r-character-method}
}
}

