% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{openRessource,character-method}
\alias{openRessource,character-method}
\title{Open Directory in File System Browser}
\usage{
\S4method{openRessource}{character}(path = ".", strict = FALSE, ...)
}
\arguments{
\item{path}{\code{\link{character}}. File path.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: non-existing ressources trigger error;
\code{FALSE}: non-existing ressources result in return value of \code{FALSE}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}: successfully opened if
   ressource exists;
   \code{FALSE}: failed to open ressource as it does not exist.
}
\description{
See generic: \code{\link[filesystr]{openRessource}}
}
\examples{
\dontrun{

## Directory //
openRessource(path = tempdir())

## File //
path <- file.path(tempdir(), "test.txt")
file.create(path)
openRessource(path = path)

## Non-existing ressource //
openRessource(path = "abdce")
try(openRessource(path = "abdce", strict = TRUE))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
   \link[filesystr]{openRessource}
}
}

