% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openRessource}
\alias{openRessource}
\title{Open Directory in File System Browser}
\usage{
openRessource(path = ".", strict = FALSE, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing directory path information.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: non-existing ressources trigger error;
\code{FALSE}: non-existing ressources result in return value of \code{FALSE}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Opens a directory in a file system browser
}
\examples{
\dontrun{

## Directory //
## Absolute path:
openRessource(path = tempdir())
## Relative path (non-existing):
openRessource(path = "does-not-exist")
## Relative path (existing):
openRessource(path = "R")

## File //
## Absolute path:
path <- file.path(tempdir(), "test.txt")
file.create(path)
openRessource(path = path)
## Relative path
openRessource(path = "DESCRIPTION")

## Non-existing ressource //
openRessource(path = "abdce")
try(openRessource(path = "abdce", strict = TRUE))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[filesystr]{openRessource-character-method}
}
}

