% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{recomposePath,RappFilesystemDecomposedPathS3-method}
\alias{recomposePath,RappFilesystemDecomposedPathS3-method}
\title{Recompose Path}
\usage{
\S4method{recomposePath}{RappFilesystemDecomposedPathS3}(path,
  shortform = FALSE, ...)
}
\arguments{
\item{path}{\code{\link{RappFilesystemDecomposedPathS3}}.
Decomposed path object as returned by
  \code{\link[filesystr]{decomposePath}}.}

\item{shortform}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means the short forms for \code{getwd()} (i.e. \code{"."},
\code{dirname(getwd()} (i.e. \code{".."} and \code{HOME} (i.e. \code{"~"})
are used, \code{FALSE} means they are not used.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. Vector of paths (directories and/or
		files).
}
\description{
See \code{\link[filesystr]{recomposePath}}.
}
\examples{
\dontrun{

 path <- c(
  "dir_b/sub_1",
  "dir_b/sub_1/",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/sub_1/notthere.txt",
  "dir_b/sub_1/notthere.txt/",
  "dir_b/noext",
  "dir_b/noext/",
  "dir_b/noext/file.txt",
  "dir_b/noext/file.txt/",
  "dir_b/noext/notthere.txt",
  "dir_b/noext/notthere.txt/"
)
path_dec <- decomposePath(path = path)
path_rec <- recomposePath(path = path_dec)
identical(path_rec, standardizePath(path))

path_rec <- recomposePath(path = path_dec, shortform = TRUE)
identical(path_rec, gsub("/$", "", path))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{recomposePath},
		\link[filesystr]{decomposePath}
}
}

