% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recomposePath}
\alias{recomposePath}
\title{Recompose Path}
\usage{
recomposePath(path, shortform = FALSE, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Recomposes path components as returned by
\code{\link[filesystr]{decomposePath}} to complete paths.
}
\examples{
\dontrun{

 path <- c(
  "dir_b/sub_1",
  "dir_b/sub_1/",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/sub_1/notthere.txt",
  "dir_b/sub_1/notthere.txt/",
  "dir_b/noext",
  "dir_b/noext/",
  "dir_b/noext/file.txt",
  "dir_b/noext/file.txt/",
  "dir_b/noext/notthere.txt",
  "dir_b/noext/notthere.txt/"
)
path_dec <- decomposePath(path = path)
path_rec <- recomposePath(path = path_dec)
identical(path_rec, standardizePath(path))

path_rec <- recomposePath(path = path_dec, shortform = TRUE)
identical(path_rec, gsub("/$", "", path))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[filesystr]{recomposePath-RappFilesystemDecomposePathS3-method},
   \link[filesystr]{decomposePath}
}
}

