% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{standardizePath,character-method}
\alias{standardizePath,character-method}
\title{Standardize Path}
\usage{
\S4method{standardizePath}{character}(path = ".", sep = c("/", "\\\\"),
  shortform = FALSE, ...)
}
\arguments{
\item{path}{\code{\link{character}}. Vector of file paths.
Default: files in current directory.}

\item{sep}{\code{\link{character}}.
String separator betwen directory levels in the output.}

\item{shortform}{\code{\link{logical}}. Dimension: 1.
Use shortforms (\code{TRUE}) for \code{getwd()} (i.e. \code{"."},
  \code{dirname(getwd()} (i.e. \code{".."} and \code{HOME} (i.e. \code{"~"})
  or not (\code{FALSE}, default).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{matrix}} with three
   columns named \code{"dirname"}, \code{"filename"} and \code{"extension"}.
   \code{strip_extension} returns a character vector of the second column,
   and \code{get_extension} returns a character vector of the third column.
\code{recompose_path} returns a character vector of paths.
}
\description{
See \code{\link[filesystr]{standardizePath}}.
}
\examples{
\dontrun{

standardizePath()

path <- c(".", "./dir", "..", "../dir", "~", "~/file.txt", R.home(), 
            tempdir(), "nonex", "nonex/", "nonex/file.txt", 
            "nonex/file.txt/", "", NA)
standardizePath(path = path, sep = "\\\\")
try(standardizePath(path = path, sep = "notallowed"))
standardizePath(path = path, shortform = TRUE)

standardizePath(".")
standardizePath(path = character())
standardizePath(path = "")
standardizePath(path = NA_character_)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{standardizePath}
}
}

