% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{standardizePath,missing-method}
\alias{standardizePath,missing-method}
\title{Standardize Path}
\usage{
\S4method{standardizePath}{missing}(path = ".", sep = c("/", "\\\\"),
  shortform = FALSE, ...)
}
\arguments{
\item{path}{\code{\link{missing}}.}

\item{sep}{\code{\link{character}}.
String separator betwen directory levels in the output.}

\item{shortform}{\code{\link{logical}}. Dimension: 1.
Use shortforms (\code{TRUE}) for \code{getwd()} (i.e. \code{"."},
  \code{dirname(getwd()} (i.e. \code{".."} and \code{HOME} (i.e. \code{"~"})
  or not (\code{FALSE}, default).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[filesystr]{standardizePath-character-method}}.
}
\description{
See \code{\link[filesystr]{standardizePath}}.
}
\examples{
\dontrun{

standardizePath()

path <- c(".", "./dir", "..", "../dir", "~", "~/file.txt", R.home(), 
            tempdir(), "nonex", "nonex/", "nonex/file.txt", 
            "nonex/file.txt/", "", NA)
standardizePath(path = path, sep = "\\\\")
try(standardizePath(path = path, sep = "notallowed"))
standardizePath(path = path, shortform = TRUE)

standardizePath(".")
standardizePath(path = character())
standardizePath(path = "")
standardizePath(path = NA_character_)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
		\link[filesystr]{standardizePath}
		\link[filesystr]{standardizePath-character-method}
}
}

