% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{stripFileExtension,missing-method}
\alias{stripFileExtension,missing-method}
\title{Get File Extension}
\usage{
\S4method{stripFileExtension}{missing}(path = ".", include_dir = TRUE, ...)
}
\arguments{
\item{path}{\code{\link{missing}}.}

\item{include_dir}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} means directory path is included, \code{FALSE} means it is
  not.}

\item{...}{Further arguments passed to subsequent functions/methods.
In particular:
\code{\link[filesystr]{decomposePath}}}
}
\value{
See method
   \code{\link[filesystr]{stripFileExtension-character-method}}
}
\description{
See generic: \code{\link[filesystr]{stripFileExtension}}.
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
stripFileExtension(path = path)
stripFileExtension(path = path, include_dir = FALSE)
stripFileExtension(path = character())
stripFileExtension()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{stripFileExtension},
		\link[filesystr]{stripFileExtension-character-method}
}
}

