% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripFileExtension}
\alias{stripFileExtension}
\title{Strip File Extension}
\usage{
stripFileExtension(path = ".", include_dir = TRUE, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing file path/file name information.}

\item{include_dir}{\code{\link{logical}}. Dimension: 1.
Inlcude directory path (\code{TRUE}, default) or not (\code{FALSE}).}

\item{...}{Further arguments passed to subsequent functions/methods.
In particular:
\code{\link[filesystr]{decomposePath}}}
}
\description{
Strips file extensions.
}
\examples{
\dontrun{

path <- c(
  ".",
  "..",
  "dir_a", 
  "dir_a/",
  "dir_b", 
  "dir_b/sub_1",
  "dir_b/sub_1/file.txt",
  "dir_b/sub_1/file.txt/",
  "dir_b/nonex",
  "dir_b/nonex/",
  "dir_b/nonex/file.txt",
  "nonex",
  "nonex/sub_1",
  "nonex/sub_1/file.txt",
  "nonex/foo.tgz", 
  "another nonex\\\\file.tar.gz", 
  "a_file", 
  "quux. quuux.tbz2", 
  "~/file.tar.xz",
  "", 
  NA_character_
)
stripFileExtension(path = path)
stripFileExtension(path = path, include_dir = FALSE)
stripFileExtension(path = character())
stripFileExtension()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp/filesystr}
}
\seealso{
\code{
		\link[filesystr]{stripFileExtension-character-method}
}
}

