% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tryUnlockRessource}
\alias{tryUnlockRessource}
\title{Try to Unlock Files System Ressource}
\usage{
tryUnlockRessource(path, try_limit = 100, strict = FALSE, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing file system path information.}

\item{try_limit}{\code{\link{numeric}}.
Limit for trying to lock ressource.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: when \code{try_limit} is reached, an error is thrown;
\code{FALSE}: when \code{try_limit} is reached, method returns with \code{FALSE}.}

\item{...}{Further arguments passed to subsequent functions/methods.
In particular:
     \code{\link[filesystr]{isRessourceLocked},
   \link[filesystr]{unlockRessource}}.}
}
\description{
Unlocks file system ressource if it is locked yet, else tries \code{try_limit}
times to unlock it. If \code{try_limt}
is reached, either \code{FALSE} is returned or an error is thrown
(depends on \code{strict}).
}
\examples{
\dontrun{

## Example content //
path <- file.path(tempdir(), "test/registry.rdata")
file.create(path)
  
## File //
tryUnlockRessource(path = path, try_limit = 3)
try(tryUnlockRessource(path = path, try_limit = 3, strict = TRUE))
lockRessource(path = path)
tryUnlockRessource(path = path)

## Directory //
path <- dirname(path)
tryUnlockRessource(path = path, try_limit = 3)
try(tryUnlockRessource(path = path, try_limit = 3, strict = TRUE))
lockRessource(path = path)
tryUnlockRessource(path = path)

## Non-existing ressource //
expect_error(tryUnlockRessource(path = "abcd", strict = TRUE))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[filesystr]{tryUnlockRessource-character-method}
}
}

