% Generated by roxygen2 (4.0.2): do not edit by hand
\name{asDirectoryPath}
\alias{asDirectoryPath}
\title{As Directory Path}
\usage{
asDirectoryPath(path = ".", ensure = FALSE, ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{ensure}{\code{logical}.
Ensure Directory existence (\code{TRUE}) or not (\code{FALSE}, default).}

\item{...}{Further arguments passed to:
\code{\link[filesystr]{ensureDirectory}}.}
}
\description{
Turns a \code{character} object denoting a directory path into an object
of class \code{RappFilesystemDirectory3}.
}
\examples{
\dontrun{

## Auxiliary function //
.cleanTempDir <- function(x) {
  if (grepl(basename(tempdir()), x)) {
    unlink(x, recursive = TRUE, force = TRUE)
  }
}

## Path //
path_0 <- file.path(tempdir(), "path")

## character //
res <- asDirectoryPath(path = path_0)
res
class(res)

## Directory.S3 //
res <- asDirectoryPath(path = asDirectoryPath(path_0))
res
class(res)

## Ensure //
file.exists(asDirectoryPath(path = path_0, ensure = TRUE))
.cleanTempDir(x = path_0)

file.exists(
  asDirectoryPath(path = asDirectoryPath(path_0), ensure = TRUE)
)
.cleanTempDir(x = path_0)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{https://github.com/Rappster/filesystr}
}
\seealso{
\code{
   \link[base]{asDirectoryPath-character-method},
		\link[filesystr]{ensureDirectory}
}
}

